LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/preload/link-header-preload-nonce-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/dummy.js?from-header&without-nonce because it does not appear in the script-src directive of the Content Security Policy.

PASS Makes sure that Link headers preload resources with CSP nonce
